check before deleting
git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.contentlibrary_rendering/trunk@108404 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/nti/contentlibrary_rendering/locators.py
@@ -120,7 +120,8 @@ class FilesystemLocator(LocatorMixin):
         return root.getChildNamed(name)
 
     def _do_remove(self, bucket):
-        shutil.rmtree(bucket.absolute_path)
+        if os.path.exists(bucket.absolute_path):
+            shutil.rmtree(bucket.absolute_path)
 
 
 @interface.implementer(IRenderedContentLocator)
